requirements.txt
cycler==0.10.0
decorator==4.2.1
Django==1.11.9
djangorestframework==3.10.1
matplotlib==2.1.2
networkx==2.1
numpy==1.14.1
opencv-python==4.1.0.25
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
